+++ PddOpenSdk/PddOpenSdk/Models/Request/Goods/CreateGoodsCpsMallUnitRequestModel.cs
@@ -7,7 +7,7 @@ namespace PddOpenSdk.Models.Request.Goods
         /// 佣金比（千分比）
         /// </summary>
         [JsonProperty("rate")]
-        public int Rate { get; set; }
+        public long Rate { get; set; }
 
     }
